requirements.txt
python-dotenv~=1.0.1
flask~=3.1.0
requests~=2.32.3
openai~=1.61.1
werkzeug~=3.1.3